pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "databricks-mcp"
version = "0.1.0"
description = "Databricks MCP server POC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.4.0",
    "databricks-sql-connector>=3.0.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["src"]
